Add beginnings of lfo, fast exp
This patch adds the low-level implementation of an LFO, with the DX7
waveforms, but doesn't yet contain the note wiring. It also adds a fast
lookup table based exp function, which is mostly used for envelopes.

There are also some build tweaks. It's possible some build files are
out of sync with the source, but at least the Android app seems to
build.