Gather more stats
Add a button to capture raw start and end callback time (just putting it
in a text buffer so it can be copied), which is useful for making plots.

Also a bunch of commented out code to provoke priority inversion or
sudden increases in load (which exercises the governor), again for the
purpose of probing performance and making plots.