Move USB device to SynthesizerService
This patch moves the connection to the USB MIDI keyboard from a single
activity to the SynthesizerService so that it can run while the user
switches between more than one activity.

Note that in the current state, there are problems when the device is
detached when the service isn't running.