More KeyboardView refinement
This patch adds text labels (the logic is there for all keys, but we
only draw on C), glissando. and some fine-tuning. The glissando is more
because some touch events are misclassified as move, rather than to
implement actual glissando.

Also, this patch adds velocity, but not yet with adjustable sensitivity.